Rename region to priority
This is a user-visible name, so requires precision.
It is unclear what the term "region" refers to. Region of what?
On the other hand, "priority" immediately tells you what it does:
high priority tasks execute before low priority tasks.
Using low values for high priority and high values for low priority is
also a well-understood convention.