Simplify Task states
This is a fix for a tricky bug.
For a parent task with at least two children,
if those children yield simultaneously,
they will both trigger the parent task.
The first succeeds: WAITING => PENDING,
but the second fails: PENDING => PENDING.

The simplest fix was to allow a task to be pending multiple times.
But that implies the task can be simultaneously RUNNING and PENDING.
This simplifies the design, and improves performance,
but I wish I could keep some of the finer-granted state checking.