Fix bug with any_event function
Previously it would wait for all events,
regardless of whether one or more is already set.
This would fail silently in most cases,
but would hang in the case where all events are set already.

Update any_event so it tests events for being set prior to the await.
Return the first event set, regardless of whether await is necessary.